# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.AllTask:
    properties:
      tasks:
        items:
          $ref: '#/definitions/models.TaskRes'
        type: array
    type: object
...
  models.FailureInfo:
    properties:
      description:
        type: string
      error: {}
      status_code:
        type: integer
    type: object
...
  models.StatusUpdate:
    properties:
      id:
        type: integer
      status:
        type: string
    type: object
  models.Success:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
  models.TaskReq:
    properties:
      description:
        type: string
      developer_id:
        type: string
      end_date:
        example: 2006/01/02
        type: string
      file_url:
        type: string
      started_date:
        example: 2006/01/02
        type: string
      title:
        type: string
    type: object
  models.TaskRes:
    properties:
      created_at:
        type: string
      description:
        type: string
      developer_id:
        type: string
      end_date:
        example: 2006/01/02
        type: string
      file_url:
        type: string
      id:
        type: integer
      started_date:
        example: 2006/01/02
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.TaskUpdateReq:
    properties:
      description:
        type: string
      end_date:
        example: 2006/01/02
        type: string
      file_url:
        type: string
      id:
        type: integer
      started_date:
        example: 2006/01/02
        type: string
      title:
        type: string
    type: object
host: localhost:7777
info:
  contact:
    email: gofurovmurtazoxon@gmail.com
    name: Murtazoxon
    url: https://t.me/murtazokhon_gofurov
  description: This is Task api server
  termsOfService: 2 term OSG
  title: Online Service Group Task api
  version: "1.0"
paths:
...
  /task:
    post:
      consumes:
      - application/json
      description: Through this api, can create task
      parameters:
      - description: CreateTask
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TaskReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TaskRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: Through this api, can update task
      parameters:
      - description: UpdateTask
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TaskUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TaskRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: UPDATE task
      tags:
      - Task
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: Through this api, can delete task
      parameters:
      - description: task_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: DELETE task
      tags:
      - Task
    get:
      consumes:
      - application/json
      description: Through this api, can create task
      parameters:
      - description: task_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TaskRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - Task
  /task/status:
    put:
      consumes:
      - application/json
      description: Through this api, can update task status
      parameters:
      - description: Update status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StatusUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusUpdate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Update task status
      tags:
      - Task
  /tasks:
    get:
      consumes:
      - application/json
      description: Through this api, can create task
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllTask'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - Task
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"